fix(db): widen summaries.version to bigint to match legacy data
Bug 3 from the live-Postgres T2 verification pass. Legacy
`BaseModel.save()` overrides `Summary.version` with the same
millisecond-resolution timestamp it uses for `server_version` (see
`app/cli/_legacy_peewee_models/_base.py:42-43`). Those values are
~1.7e12 — they fit SQLite's flexible INTEGER but overflow Postgres
INTEGER (max 2.1e9), so the SQLite -> Postgres migrator failed on the
first Summary insert with `asyncpg.DataError: value out of int32
range`. Production-relevant: the live Pi `summaries` table has the
same value pattern, so this would block the C2 cutover.

Changes:

- `app/db/models/core.py:373` — `Summary.version` widened from
  `Integer` to `BigInteger`. Now matches the existing `server_version`
  type and the legacy ms-timestamp value range.
- `app/db/alembic/versions/0002_widen_summary_version_to_bigint.py` —
  new revision `0002` (down=`0001`) that issues `ALTER COLUMN ...
  TYPE BIGINT` on the live Postgres. Forward + downgrade both
  defined.
- `app/cli/migrate_sqlite_to_postgres.py` — extended the
  `_legacy_table_exists` guard from commit `4291f16e` to
  `_sample_fk_cardinalities` (was missed in that commit; without it
  validation crashed when the source SQLite had only a partial
  schema).
- `tests/cli/test_migrate_sqlite_to_postgres.py` — added a
  `_clear_target` step at the start of `test_bytes_column_round_trip`
  that TRUNCATEs all tables before the migration runs. T2 tests
  bypass the conftest `session` fixture (they call `run_migration`
  directly), so without explicit cleanup they would inherit rows from
  earlier tests in the same session and validation would fail with a
  spurious source/target count mismatch.

Verification (live Postgres on port 5433):

- `alembic upgrade head` applies cleanly: `Running upgrade 0001 ->
  0002, widen summaries.version from INTEGER to BIGINT.`
- `\d summaries` confirms `version | bigint`.
- Combined run of T2 + T3 smoke suites: **16 of 16 passing** (was
  15/16 with `test_bytes_column_round_trip` failing).
- ruff + mypy clean across all touched files.

+++ app/db/models/core.py
@@ -370,7 +370,7 @@ class Summary(Base):
     lang: Mapped[str | None] = mapped_column(Text, nullable=True)
     json_payload: Mapped[JSONValue] = _json_column()
     insights_json: Mapped[JSONValue] = _json_column()
-    version: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
+    version: Mapped[int] = mapped_column(BigInteger, default=1, nullable=False)
     server_version: Mapped[int] = mapped_column(
         BigInteger, default=_next_server_version, nullable=False
     )
